requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
anyio==4.10.0
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
emoji==2.14.1
exceptiongroup==1.3.0
faiss-cpu==1.7.4
fastapi==0.100.0
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.7.0
ftfy==6.1.1
gitdb==4.0.12
GitPython==3.1.45
h11==0.14.0
hf-xet==1.1.7
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.2
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langchain==0.1.16
langchain-community==0.0.38
langchain-core==0.1.53
langchain-text-splitters==0.0.2
langgraph==0.0.34
langsmith==0.1.147
loguru==0.7.2
lxml==6.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==2.0.1
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
openai==1.30.1
orjson==3.11.1
packaging==23.2
pandas==2.3.1
pillow==10.2.0
propcache==0.3.2
protobuf==6.31.1
pyarrow==21.0.0
pydantic==1.10.13
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2023.12.25
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.27.0
safetensors==0.6.2
scipy==1.13.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.42
starlette==0.27.0
streamlit==1.48.0
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.5.2
tokenizers==0.15.2
toml==0.10.2
torch==2.2.2
tornado==6.5.2
tqdm==4.67.1
transformers==4.39.3
typing-inspect==0.9.0
typing_extensions==4.14.1
tzdata==2025.2
unstructured==0.10.16
urllib3==2.5.0
uvicorn==0.23.2
watchdog==6.0.0
wcwidth==0.2.13
yarl==1.20.1
